refactor: 정의한 컴포넌트와 파일 이름 같게 수정
