+++ src/contexts/index.ts
@@ -1,1 +1,1 @@
-export * from './modalContexts';
+export * from './ModalContext';
